Move test helper devices to headers
