+++ test/helper/devices/TestPollIn.h
@@ -6,6 +6,9 @@
  *                  copied verbatim in the file "LICENSE"                       *
  ********************************************************************************/
 
+#ifndef FAIR_MQ_TEST_POLLIN_H
+#define FAIR_MQ_TEST_POLLIN_H
+
 #include <FairMQDevice.h>
 #include <FairMQLogger.h>
 #include <options/FairMQProgOptions.h>
@@ -127,3 +130,5 @@ class PollIn : public FairMQDevice
 } // namespace test
 } // namespace mq
 } // namespace fair
+
+#endif /* FAIR_MQ_TEST_POLLIN_H */
