+++ test/helper/devices/TestRep.h
@@ -6,7 +6,11 @@
  *                  copied verbatim in the file "LICENSE"                       *
  ********************************************************************************/
 
+#ifndef FAIR_MQ_TEST_REP_H
+#define FAIR_MQ_TEST_REP_H
+
 #include <FairMQDevice.h>
+#include <FairMQLogger.h>
 #include <thread>
 
 namespace fair
@@ -16,7 +20,7 @@ namespace mq
 namespace test
 {
 
-class Push : public FairMQDevice
+class Rep : public FairMQDevice
 {
   protected:
     auto Init() -> void override
@@ -26,11 +30,27 @@ class Push : public FairMQDevice
 
     auto Run() -> void override
     {
-        auto msg = FairMQMessagePtr{NewMessage()};
-        Send(msg, "data");
+        auto request1 = FairMQMessagePtr{NewMessage()};
+        if (Receive(request1, "data") >= 0)
+        {
+            LOG(info) << "Received request 1";
+            auto reply = FairMQMessagePtr{NewMessage()};
+            Send(reply, "data");
+        }
+        auto request2 = FairMQMessagePtr{NewMessage()};
+        if (Receive(request2, "data") >= 0)
+        {
+            LOG(info) << "Received request 2";
+            auto reply = FairMQMessagePtr{NewMessage()};
+            Send(reply, "data");
+        }
+
+        LOG(info) << "REQ-REP test successfull";
     };
 };
 
 } // namespace test
 } // namespace mq
 } // namespace fair
+
+#endif /* FAIR_MQ_TEST_REP_H */
