+++ test/helper/devices/TestReq.h
@@ -6,10 +6,12 @@
  *                  copied verbatim in the file "LICENSE"                       *
  ********************************************************************************/
 
+#ifndef FAIR_MQ_TEST_REQ_H
+#define FAIR_MQ_TEST_REQ_H
+
 #include <FairMQDevice.h>
 #include <FairMQLogger.h>
-
-#include <iostream>
+#include <thread>
 
 namespace fair
 {
@@ -18,16 +20,29 @@ namespace mq
 namespace test
 {
 
-class TestWaitFor : public FairMQDevice
+class Req : public FairMQDevice
 {
-  public:
-    void Run()
+  protected:
+    auto Init() -> void override
     {
-        std::cout << "hello" << std::endl;
-        WaitFor(std::chrono::seconds(60));
+        std::this_thread::sleep_for(std::chrono::milliseconds(200));
     }
+
+    auto Run() -> void override
+    {
+        auto request = FairMQMessagePtr{NewMessage()};
+        Send(request, "data");
+
+        auto reply = FairMQMessagePtr{NewMessage()};
+        if (Receive(reply, "data") >= 0)
+        {
+            LOG(info) << "received reply";
+        }
+    };
 };
 
 } // namespace test
 } // namespace mq
 } // namespace fair
+
+#endif /* FAIR_MQ_TEST_REQ_H */
