+++ test/helper/devices/TestSub.h
@@ -6,6 +6,9 @@
  *                  copied verbatim in the file "LICENSE"                       *
  ********************************************************************************/
 
+#ifndef FAIR_MQ_TEST_SUB_H
+#define FAIR_MQ_TEST_SUB_H
+
 #include <FairMQDevice.h>
 #include <FairMQLogger.h>
 #include <chrono>
@@ -64,3 +67,5 @@ class Sub : public FairMQDevice
 } // namespace test
 } // namespace mq
 } // namespace fair
+
+#endif /* FAIR_MQ_TEST_SUB_H */
